[+] Added adaptive app icon. Fixed design issues. Improved logic of notifications manager. Fixed bugs and added improvements.

+++ app/src/main/java/com/lobstr/stellar/vault/domain/fcm/FcmInteractor.kt
@@ -13,9 +13,9 @@ interface FcmInteractor {
 
     fun getFcmToken(): String?
 
-    fun saveAppVersion(appVersion: Int)
+    fun saveAppVersion(appVersion: Long)
 
-    fun getAppVersion(): Int
+    fun getAppVersion(): Long
 
     fun setFcmRegistered(registered: Boolean)
